+++ src/classes/ValueCollection.ts
@@ -1,5 +1,7 @@
 import { RealAny } from "@/types/internal"
 
+const inspectSymbol = Symbol.for('nodejs.util.inspect.custom')
+
 export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> {
 	protected modifyFn: ((event: 'set' | 'delete' | 'clear', key: any, value: any) => any) | null = null
 	protected data: Map<Key, Value> = new Map()
@@ -21,22 +23,22 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	 * collection.has('key') // true
 	 * collection.has('ms') // false
 	 * 
-	 * collection.toJSON() // { name: 'beta', key: 'value' }
+	 * collection.json() // { name: 'beta', key: 'value' }
 	 * 
-	 * collection.forEach((key, value) => {
+	 * for (const [ key, value ] of collection) {
 	 *   console.log(key, value)
-	 * })
+	 * }
 	 * 
 	 * collection.clear(['key'])
 	 * 
-	 * collection.toJSON() // { key: 'value' }
+	 * collection.json() // { key: 'value' }
 	 * ```
 	 * @since 2.5.0
 	*/ constructor(
-		/** JSON Data to Import */ data?: Record<Key, Value>,
-		/** Function to Parse Values with */ parse?: (value: any) => Value,
-		/** Whether to allow modifying the Values */ allowModify: boolean = true,
-		/** How many Elements to store at max, when hit clear */ maxElements: number = Infinity
+		data?: Record<Key, Value>,
+		parse?: (value: any) => Value,
+		allowModify: boolean = true,
+		maxElements: number = Infinity
 	) {
 		data = data ?? {} as any
 		parse = parse ?? ((value) => value)
@@ -52,30 +54,45 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Check if a Key exists
 	 * @since 2.5.0
-	*/ public has(
-		/** The Key to check */ key: Key
-	): boolean {
+	*/ public has(key: Key): boolean {
 		return this.data.has(key)
 	}
 
 	/**
 	 * Get a Key
 	 * @since 2.5.0
-	*/ public get<T extends Key, Fallback extends Value | undefined = undefined>(
-		/** The Key to get */ key: T,
-		/** The Fallback Value */ fallback?: Fallback
-	): Value | Fallback {
+	*/ public get<Fallback extends Value | null = null>(key: Key, fallback?: Fallback): Value | Fallback {
 		return this.data.get(key) ?? (fallback as any)
+	}
+
+	/**
+	 * Search for a Key
+	 * @since 9.3.0
+	*/ public search(
+		callback: (key: Key, value: Value, index: number) => boolean
+	): [Key, Value] | null {
+		let index = 0
+		for (const [ key, value ] of this.data) {
+			if (callback(key, value, index++)) return [ key, value ]
+		}
+
+		return null
 	}
 
 	/**
 	 * Get all Objects as JSON
 	 * @since 2.5.0
-	*/ public toJSON(
-		/** Excluded Keys */ excluded: Key[] = []
-	): Record<Key, Value> {
+	 * @deprecated use `.json()` instead
+	*/ public toJSON(excluded: Key[] = []): Record<Key, Value> {
+		return this.json(excluded)
+	}
+
+	/**
+	 * Get the Data of this Collection as JSON
+	 * @since 9.3.0
+	*/ public json(excluded: Key[] = []): Record<Key, Value> {
 		let keys = {} as any
-		for (const [key, value] of this.data) {
+		for (const [ key, value ] of this.data) {
 			if (excluded.includes(key)) continue
 			keys[key] = value
 		}
@@ -86,11 +103,17 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Get all Values as Array
 	 * @since 2.5.0
-	*/ public toArray(
-		/** Excluded Keys */ excluded: Key[] = []
-	): Value[] {
-		const values = []
-		for (const [key, value] of this.data) {
+	 * @deprecated use `.values()` instead
+	*/ public toArray(excluded: Key[] = []): Value[] {
+		return this.values(excluded)
+	}
+
+	/**
+	 * Get the Values of this Collection
+	 * @since 9.3.0
+	*/ public values(excluded: Key[] = []): Value[] {
+		const values: Value[] = []
+		for (const [ key, value ] of this.data) {
 			if (excluded.includes(key)) continue
 			values.push(value)
 		}
@@ -99,12 +122,22 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	}
 
 	/**
+	 * Get the Keys of this Collection
+	 * @since 9.3.0
+	*/ public keys(excluded: Key[] = []): Key[] {
+		const keys: Key[] = []
+		for (const [ key ] of this.data) {
+			if (excluded.includes(key)) continue
+			keys.push(key)
+		}
+
+		return keys
+	}
+
+	/**
 	 * Loop over all Keys
 	 * @since 2.5.0
-	*/ public forEach(
-		/** Callback Function */ callback: (key: Key, value: Value, index: number) => RealAny,
-		/** Excluded Keys */ excluded: Key[] = []
-	): this {
+	*/ public forEach(callback: (key: Key, value: Value, index: number) => RealAny, excluded: Key[] = []): this {
 		callback = callback ?? (() => undefined)
 
 		let index = 0
@@ -125,11 +158,9 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	}
 
 	/**
-	 * Get the Entries of this Value Collection
+	 * Get the Entries of this Collection
 	 * @since 6.0.3
-	*/ public entries(
-		/** Excluded Keys */ excluded: Key[] = []
-	): [Key, Value][] {
+	*/ public entries(excluded: Key[] = []): [Key, Value][] {
 		const entries: [Key, Value][] = []
 		
 		this.data.forEach((value, key) => {
@@ -145,8 +176,8 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	 * Map the Keys to a new Array
 	 * @since 5.3.1
 	*/ public map<Callback extends (key: Key, value: Value, index: number, data: this) => RealAny>(
-		/** Callback Function */ callback: Callback,
-		/** Excluded Keys */ excluded: Key[] = []
+		callback: Callback,
+		excluded: Key[] = []
 	): ReturnType<Callback>[] {
 		callback = callback ?? ((value) => value)
 		const result: ReturnType<Callback>[] = []
@@ -164,8 +195,20 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * The Amount of Stored Objects
 	 * @since 2.7.2
+	 * @deprecated use `.size()` instead
 	*/ public get objectCount(): number {
 		return this.data.size
+	}
+
+	/**
+	 * Get the Size of this Collection
+	 * @since 9.3.0
+	*/ public size(): number {
+		return this.data.size
+	}
+
+	[inspectSymbol](): string {
+		return `<ValueCollection (${this.data.size})>`
 	}
 }
 
@@ -180,13 +223,10 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Set a Key
 	 * @since 2.5.0
-	*/ public set(
-		/** The Key to set */ key: Key,
-		/** The new Value */ value: SetValue
-	): this {
+	*/ public set(key: Key, value: SetValue): this {
 		if (!this.allowModify) return this
 
-		if (this.objectCount > this.maxElements) this.clear()
+		if (this.data.size > this.maxElements) this.clear()
 
 		if (this.modifyFn) this.modifyFn('set', key, value)
 		else this.data.set(key, value as any)
@@ -197,12 +237,10 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Delete a Key
 	 * @since 8.0.0
-	*/ public delete(
-		/** The Key to delete */ key: Key
-	): this {
+	*/ public delete(key: Key): this {
 		if (!this.allowModify) return this
 
-		if (this.objectCount > this.maxElements) this.clear()
+		if (this.data.size > this.maxElements) this.clear()
 
 		if (this.modifyFn) this.modifyFn('delete', key, null)
 		else this.data.delete(key)
@@ -213,9 +251,7 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Clear the Stored Objects
 	 * @since 3.0.0
-	*/ public clear(
-		/** Excluded Keys */ excluded: Key[] = []
-	): number {
+	*/ public clear(excluded: Key[] = []): number {
 		if (!this.allowModify) return 0
 
 		let keys = 0
@@ -232,10 +268,12 @@ export class BaseCollection<Key extends PropertyKey = PropertyKey, Value = any> 
 	/**
 	 * Import another Value Collection into this one
 	 * @since 9.0.0
-	*/ public import(valueCollection: ValueCollection<Key, Value, any>): this {
-		valueCollection.data.forEach((value, key) => {
-			this.data.set(key, value)
-		})
+	*/ public import(data: ValueCollection<Key, SetValue, any> | Map<Key, SetValue> | IterableIterator<[Key, SetValue]>): this {
+		if (!this.allowModify) return this
+
+		for (const [ key, value ] of data) {
+			this.set(key, value)
+		}
 
 		return this
 	}